Make service question summary stateless
