+++ src/components/Admin.js
@@ -9,6 +9,7 @@ import Users from './Users';
 import UserDetails from './UserDetails';
 import FAQAnswers from "./FAQAnswers";
 import ServiceQuestions from './ServiceQuestions';
+import ServiceQuestionsContainer from './ServiceQuestionsContainer';
 import ServiceQuestionDetails from './ServiceQuestionDetails';
 import ServiceAnswers from './ServiceAnswers';
 import ServiceAnswerDetails from './ServiceAnswerDetails';
@@ -42,7 +43,7 @@ const Admin = () => (
                     <Route path="/admin/users" exact component={Users} />
                     <Route path="/admin/users/:id" exact component={UserDetails}/>
                     <Route path="/admin/faqs/:id" exact component={FAQDetailsContainer} />
-                    <Route path="/admin/service-questions" exact component={ServiceQuestions} />
+                    <Route path="/admin/service-questions" exact component={ServiceQuestionsContainer} />
                     <Route path="/admin/service-questions/:id" exact component={ServiceQuestionDetails} />
                     <Route path="/admin/service-answers" exact component={ServiceAnswers} />
                     <Route path="/admin/service-answers/:id" exact component={ServiceAnswerDetails} />
